hr-ai-platform/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
langchain>=0.3.0,<1.0.0
langchain-nvidia-ai-endpoints>=0.3.0,<1.0.0
langgraph>=0.2.0,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
SQLAlchemy>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.0,<2.0.0
bcrypt>=4.0.0,<5.0.0
